refactor lark-string to contain a new `Text` type
